Added Base Classes to remove redundant code
