+++ lib/datapoints.js
@@ -234,7 +234,6 @@ function getDeviceByType(type, protocol) {
 
 function getDeviceNameForMQTT(name) {
   return devicenames[name] ? devicenames[name][0] : undefined;
-  // return devicenames[name];
 }
 
 function getDeviceNameForCoAP(name) {
