+++ lib/protocol/coap.js
@@ -5,12 +5,13 @@
 /* jslint esversion: 6 */
 'use strict';
 
+const BaseClient = require(__dirname + '/base').BaseClient;
+const BaseServer = require(__dirname + '/base').BaseServer;
+const datapoints = require(__dirname + '/../datapoints');
 
 const Shelly = require('shelly-iot');
-const ping = require('ping');
 const tcpp = require('tcp-ping');
 const request = require('request');
-const datapoints = require(__dirname + '/datapoints');
 
 function sleep(ms) {
   return new Promise((resolve) => setTimeout(resolve, ms));
@@ -19,22 +20,6 @@ function sleep(ms) {
 function isAsync(funct) {
   if (funct && funct.constructor) return funct.constructor.name == 'AsyncFunction';
   return undefined;
-}
-
-/**
- * ping host
- * @param {string} host - hostname or ip-address
- */
-function pingAsyncOld(host) {
-  return new Promise((resolve, reject) => {
-    ping.sys.probe(host, (isAlive) => {
-      if (isAlive) {
-        resolve(true);
-      } else {
-        resolve(false);
-      }
-    });
-  });
 }
 
 /**
@@ -123,31 +108,20 @@ function descrToSensor(description) {
   return sensors;
 }
 
-class CoAPClient {
+class CoAPClient extends BaseClient {
   constructor(adapter, objectHelper, eventEmitter, shelly, devicename, ip, payload, description) {
+
+    super(adapter, objectHelper, eventEmitter);
+
     this.getOldDeivceInfo(devicename);
-    this.active = true;
-    this.adapter = adapter;
-    this.objectHelper = objectHelper;
-    this.eventEmitter = eventEmitter;
+
     this.shelly = shelly;
     this.devicename = devicename;
     this.ip = ip;
-    this.states = {};
-    this.device = {};
-    this.http = {};
-    this.auth;
-    this.id;
-    this.devicetype;
-    this.deviceid;
-    this.serialid;
-    this.deviceexist;
-    this.devicegen;
-    this.httptimeout = 5 * 1000;
+
     this.description = description;
     this.sensors = descrToSensor(description);
-    if (this.adapter.config.httpusername && this.adapter.config.httppassword && this.adapter.config.httpusername.length > 0 && this.adapter.config.httppassword.length > 0)
-      this.auth = 'Basic ' + Buffer.from(this.adapter.config.httpusername + ':' + this.adapter.config.httppassword).toString('base64');
+
     this.start(payload, description);
   }
 
@@ -220,41 +194,6 @@ class CoAPClient {
   }
 
   /**
-   * get the Polltime for a device
-   */
-  getPolltime() {
-    if (this.adapter.config.polltime == 0) return 0;
-    let deviceid = this.getDeviceId();
-    let polltime = undefined;
-    if (deviceid) {
-      polltime = datapoints.getPolltime(deviceid);
-      if (polltime === undefined) polltime = this.adapter.config.polltime || 0;
-      if (this.adapter.config.polltime > polltime) polltime = this.adapter.config.polltime;
-      return polltime;
-    }
-    return this.adapter.config.polltime;
-  }
-
-  async onlineCheck(stop) {
-    let deviceId = this.getDeviceName();
-    let polltime = this.getPolltime();
-    clearTimeout(this.onlinehandle);
-    try {
-      if (stop === 'stop') {
-        await this.adapter.setStateAsync(deviceId + '.online', {val: false, ack: true});
-      } else {
-        this.onlinehandle = setTimeout(async () => {
-          await this.adapter.setStateAsync(deviceId + '.online', {val: false, ack: true});
-        }, (polltime * 1000) + 10 * 1000);
-        let state = await this.adapter.getStateAsync(deviceId + '.online');
-        if (state && state.val === false) await this.adapter.setStateAsync(deviceId + '.online', {val: true, ack: true});
-      }
-    } catch (err) {
-      // Ignore
-    }
-  }
-
-  /**
    * Get the serialid back without devicename.
    * For example, you get 12345 for shellyplug-s-12345 back
    */
@@ -270,66 +209,18 @@ class CoAPClient {
   }
 
   /**
-   * Checks if Shelly device type in the configuration exist. If missing
-   * you have to add a configuration in the ./lib/devices direcotory
-   */
-  deviceExist() {
-    if (this.deviceexist === undefined) {
-      let deviceid = this.getDeviceId();
-      this.deviceexist = datapoints.getDeviceNameForMQTT(deviceid) ? true : false;
-    }
-    return this.deviceexist;
-  }
-
-  /**
-    * Checks if Shelly supports rpc api
-    * you have to add a configuration in the ./lib/devices direcotory
-    */
-  getDeviceGen() {
-    if (this.devicegen === undefined) {
-      let deviceid = this.getDeviceId();
-      this.devicegen = datapoints.getDeviceGen(deviceid);
-      this.adapter.log.debug('DEVICE GEN: ' + this.devicegen);
-    }
-    return this.devicegen;
-  }
-
-  /**
-   * Returns a string for Logging with the IP address and name of Shelly Device and type
-   */
-  getName() {
-    let name = this.getDeviceName(); // SHRGBW2#1234#1
-    let ip = this.getIP(); // 192.168.11.1
-    let deviceid = this.getDeviceId(); // shellyplug-s-12345
-    let id = this.getId(); // shellyplug-s-12345
-    return ip + ' (' + deviceid + ' / ' + id + ' / ' + name + ')';
-  }
-
-  /**
    * Cleanup, destroy this object
    */
   destroy() {
-    if (this.active) {
-      this.adapter.log.info('Destroy ' + this.getName());
-      this.active = false;
-      clearTimeout(this.timerid);
-      clearTimeout(this.autoupdateid);
-      this.devicename = undefined;
-      this.http = {};
-      this.states = {};
-      this.device = {};
-      this.ip = undefined;
-      this.id = undefined;
-      this.devicetype = undefined;
-      this.deviceid = undefined;
-      this.serialid = undefined;
-      this.deviceexist = undefined;
-      this.devicegen = undefined;
-      this.description = undefined;
-      if (this.listenerus) this.shelly.removeListener('update-device-status', this.listenerus);
-      if (this.listenerds) this.shelly.removeListener('device-connection-status', this.listenerds);
-      this.onlineCheck('stop');
-    }
+    super.destroy();
+
+    clearTimeout(this.timerid);
+    clearTimeout(this.autoupdateid);
+
+    this.description = undefined;
+
+    if (this.listenerus) this.shelly.removeListener('update-device-status', this.listenerus);
+    if (this.listenerds) this.shelly.removeListener('device-connection-status', this.listenerds);
   }
 
   /**
@@ -545,7 +436,6 @@ class CoAPClient {
       }
     }
     this.objectHelper.processObjectQueue(() => { });
-    await this.onlineCheck();
   }
 
   /**
@@ -626,7 +516,6 @@ class CoAPClient {
     if (this.http && Object.keys(this.http).length > 0 && polltime > 0) {
       this.timerid = setTimeout(async () => await this.httpIoBrokerState(), polltime * 1000);
     }
-    await this.onlineCheck();
   }
 
   /*
@@ -709,18 +598,17 @@ class CoAPClient {
     });
   }
 
-
 }
 
-class CoAPServer {
+class CoAPServer extends BaseServer {
 
   constructor(adapter, objectHelper, eventEmitter) {
-    if (!(this instanceof CoAPServer)) return new CoAPServer(adapter, objectHelper, eventEmitter);
-    this.adapter = adapter;
-    this.objectHelper = objectHelper;
+    //if (!(this instanceof CoAPServer)) return new CoAPServer(adapter, objectHelper, eventEmitter);
+
+    super(adapter, objectHelper, eventEmitter);
+
     this.clients = {};
     this.blacklist = {};
-    this.eventEmitter = eventEmitter;
   }
 
   isBlackListed(deviceId) {
