+++ lib/protocol/mqtt.js
@@ -5,13 +5,14 @@
 /* jslint esversion: 6 */
 'use strict';
 
+const BaseClient = require(__dirname + '/base').BaseClient;
+const BaseServer = require(__dirname + '/base').BaseServer;
+const datapoints = require(__dirname + '/../datapoints');
 
 const mqtt = require('mqtt-connection');
-const ping = require('ping');
 const tcpp = require('tcp-ping');
 const net = require('net');
 const request = require('request');
-const datapoints = require(__dirname + '/datapoints');
 
 /**
  * wait x miliseconds
@@ -28,22 +29,6 @@ function sleep(ms) {
 function isAsync(funct) {
   if (funct && funct.constructor) return funct.constructor.name == 'AsyncFunction';
   return undefined;
-}
-
-/**
- * ping host
- * @param {string} host - hostname or ip-address
- */
-function pingAsyncOld(host) {
-  return new Promise((resolve, reject) => {
-    ping.sys.probe(host, (isAlive) => {
-      if (isAlive) {
-        resolve(true);
-      } else {
-        resolve(false);
-      }
-    });
-  });
 }
 
 /**
@@ -102,34 +87,21 @@ function recursiveSubStringReplace(source, pattern, replacement) {
   return recursiveReplace(source);
 }
 
-class MQTTClient {
+class MQTTClient extends BaseClient {
   constructor(adapter, objectHelper, eventEmitter, stream) {
-    this.active = true;
-    this.adapter = adapter;
-    this.objectHelper = objectHelper;
-    this.eventEmitter = eventEmitter;
+
+    super(adapter, objectHelper, eventEmitter);
+
+    this.stream = stream;
+
     this.packet;
+    this.mqttprefix;
     this.qos = {};
     this.messageIds = [];
     this.messageId = 1;
-    this.states = {};
-    this.stream = stream;
+
     this.client;
-    this.device = {};
-    this.http = {};
-    this.auth;
-    this.ip;
-    this.id;
-    this.devicename;
-    this.devicetype;
-    this.deviceid;
-    this.serialid;
-    this.mqttprefix;
-    this.deviceexist;
-    this.devicegen;
-    this.httptimeout = 5 * 1000;
-    if (this.adapter.config.httpusername && this.adapter.config.httppassword && this.adapter.config.httpusername.length > 0 && this.adapter.config.httppassword.length > 0)
-      this.auth = 'Basic ' + Buffer.from(this.adapter.config.httpusername + ':' + this.adapter.config.httppassword).toString('base64');
+
     this.start();
   }
 
@@ -209,41 +181,6 @@ class MQTTClient {
   }
 
   /**
-   * get the Polltime for a device
-   */
-  getPolltime() {
-    if (this.adapter.config.polltime == 0) return 0;
-    let deviceid = this.getDeviceId();
-    let polltime = undefined;
-    if (deviceid) {
-      polltime = datapoints.getPolltime(deviceid);
-      if (polltime === undefined) polltime = this.adapter.config.polltime || 0;
-      if (this.adapter.config.polltime > polltime) polltime = this.adapter.config.polltime;
-      return polltime;
-    }
-    return this.adapter.config.polltime;
-  }
-
-  async onlineCheck(stop) {
-    let deviceId = this.getDeviceName();
-    let polltime = this.getPolltime();
-    clearTimeout(this.onlinehandle);
-    try {
-      if (stop === 'stop') {
-        await this.adapter.setStateAsync(deviceId + '.online', {val: false, ack: true});
-      } else {
-        this.onlinehandle = setTimeout(async () => {
-          await this.adapter.setStateAsync(deviceId + '.online', {val: false, ack: true});
-        }, (polltime * 1000) + 10 * 1000);
-        let state = await this.adapter.getStateAsync(deviceId + '.online');
-        if (state && state.val === false) await this.adapter.setStateAsync(deviceId + '.online', {val: true, ack: true});
-      }
-    } catch (err) {
-      // Ignore
-    }
-  }
-
-  /**
    * Get the serialid back without devicename.
    * For example, you get 12345 for shellyplug-s-12345 back
    */
@@ -259,72 +196,25 @@ class MQTTClient {
   }
 
   /**
-    * Checks if Shelly device type in the configuration exist. If missing
-    * you have to add a configuration in the ./lib/devices direcotory
-    */
-  deviceExist() {
-    if (this.deviceexist === undefined) {
-      let deviceid = this.getDeviceId();
-      this.deviceexist = datapoints.getDeviceNameForMQTT(deviceid) ? true : false;
-    }
-    return this.deviceexist;
-  }
-
-  /**
-    * Checks if Shelly supports rpc api
-    * you have to add a configuration in the ./lib/devices directory
-    */
-  getDeviceGen() {
-    if (this.devicegen === undefined) {
-      let deviceid = this.getDeviceId();
-      this.devicegen = datapoints.getDeviceGen(deviceid);
-      this.adapter.log.debug('DEVICE GEN: ' + this.devicegen);
-    }
-    return this.devicegen;
-  }
-
-  /**
-   * Returns a string for Logging with the IP address and name of Shelly Device and type
-   */
-  getName() {
-    let name = this.getDeviceName(); // SHRGBW2#1234#1
-    let ip = this.getIP(); // 192.168.11.1
-    let deviceid = this.getDeviceId(); // shellyplug-s-12345
-    let id = this.getId(); // shellyplug-s-12345
-    return ip + ' (' + deviceid + ' / ' + id + ' / ' + name + ')';
-  }
-
-  /**
    * Cleanup, destroy this object
    */
   destroy() {
-    if (this.active) {
-      this.adapter.log.debug('Destroy ' + this.getName());
-      this.active = false;
-      clearTimeout(this.timerid);
-      clearTimeout(this.autoupdateid);
-      for (let messageId in this.qos) {
-        if (this.qos[messageId].resendid) clearTimeout(this.qos[messageId].resendid);
-      }
-      this.qos = {};
-      this.messageId = 1;
-      this.states = {};
-      this.device = {};
-      this.http = {};
-      this.ip = undefined;
-      this.id = undefined;
-      this.devicename = undefined;
-      this.devicetype = undefined;
-      this.deviceid = undefined;
-      this.serialid = undefined;
-      this.deviceexist = undefined;
-      this.devicegen = undefined;
-      this.mqttprefix = undefined;
-      this.onlineCheck('stop');
-      if (this.client) {
-        this.client.removeAllListeners();
-        this.client.destroy();
-      }
+    super.destroy();
+
+    clearTimeout(this.timerid);
+    clearTimeout(this.autoupdateid);
+
+    for (let messageId in this.qos) {
+      if (this.qos[messageId].resendid) clearTimeout(this.qos[messageId].resendid);
+    }
+
+    this.qos = {};
+    this.messageId = 1;
+    this.mqttprefix = undefined;
+
+    if (this.client) {
+      this.client.removeAllListeners();
+      this.client.destroy();
     }
   }
 
@@ -640,7 +530,6 @@ class MQTTClient {
       }
     }
     this.objectHelper.processObjectQueue(() => { });
-    await this.onlineCheck();
   }
 
   /**
@@ -721,7 +610,6 @@ class MQTTClient {
     if (this.http && Object.keys(this.http).length > 0 && polltime > 0) {
       this.timerid = setTimeout(async () => await this.httpIoBrokerState(), polltime * 1000);
     }
-    await this.onlineCheck();
   }
 
   /*
@@ -989,16 +877,16 @@ class MQTTClient {
 
 }
 
-class MQTTServer {
+class MQTTServer extends BaseServer {
 
   constructor(adapter, objectHelper, eventEmitter) {
-    if (!(this instanceof MQTTServer)) return new MQTTServer(adapter, objectHelper, eventEmitter);
+    //if (!(this instanceof MQTTServer)) return new MQTTServer(adapter, objectHelper, eventEmitter);
+
+    super(adapter, objectHelper, eventEmitter);
+
     this.messageId = 1;
     this.server = new net.Server();
-    this.adapter = adapter;
     this.clients = [];
-    this.objectHelper = objectHelper;
-    this.eventEmitter = eventEmitter;
   }
 
   listen() {
